Demo commit

+++ parent/ob-common/src/main/java/by/it/academy/ob/dao/DAO.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface DAO<E> {
-    int add(E e) throws SQLException;
+    Long add(E e) throws SQLException;
     Optional<E> read(Long id) throws SQLException;
     int update(E e) throws SQLException;
-    Long delete(Long id) throws SQLException;
+    int delete(Long id) throws SQLException;
     List<E> getAll() throws SQLException;
 }
